Rename to OperationStats and make public
