+++ vespa-feed-client/src/main/java/ai/vespa/feed/client/HttpFeedClient.java
@@ -64,6 +64,11 @@ class HttpFeedClient implements FeedClient {
     }
 
     @Override
+    public OperationStats stats() {
+        return requestStrategy.stats();
+    }
+
+    @Override
     public void close(boolean graceful) {
         closed.set(true);
         if (graceful)
