+++ vespa-feed-client/src/test/java/ai/vespa/feed/client/JsonFeederTest.java
@@ -72,6 +72,9 @@ class JsonFeederTest {
                 }
 
                 @Override
+                public OperationStats stats() { return null; }
+
+                @Override
                 public void close(boolean graceful) { }
 
                 private CompletableFuture<Result> createSuccessResult(DocumentId documentId) {
